Database updates
- Database seeds to fill in tables for:
— Councils
— Postcodes
— Suburbs

— Update to postcodes table to use integer instead of smallint

+++ database/migrations/2017_05_14_0002_create_postcodes_table.php
@@ -16,7 +16,7 @@ class CreatePostcodesTable extends Migration
         // Create postcode table
         Schema::create('postcodes', function (Blueprint $table) {
             $table->increments('id')->unique();
-            $table->smallInteger('postcode');
+            $table->integer('postcode');
             $table->timestamps();
         });
     }
